backend/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cloudinary==1.44.1
Django==6.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.2.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-decouple==3.8
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
